+++ qctrlopencontrols/base/__init__.py
@@ -11,13 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-===========
-Base module
-===========
-"""
-
-from .utils import create_repr_from_attributes
-
-__all__ = ['create_repr_from_attributes']
