+++ qctrlopencontrols/cirq/schedule.py
@@ -22,8 +22,8 @@ import numpy as np
 
 import cirq
 
-from qctrlopencontrols.dynamic_decoupling_sequences import DynamicDecouplingSequence
-from qctrlopencontrols.exceptions import ArgumentsValueError
+from ..dynamic_decoupling_sequences.dynamic_decoupling_sequence import DynamicDecouplingSequence
+from ..exceptions.exceptions import ArgumentsValueError
 
 
 def convert_dds_to_cirq_schedule(
