+++ qctrlopencontrols/driven_controls/__init__.py
@@ -13,39 +13,68 @@
 # limitations under the License.
 
 """
-======================
-driven_controls module
-======================
+===============
+driven_controls
+===============
 """
 
-from .driven_control import DrivenControl
+##### Maximum and Minimum bounds ######
 
-from .constants import (
-    UPPER_BOUND_RABI_RATE, UPPER_BOUND_DETUNING_RATE,
-    UPPER_BOUND_DURATION, LOWER_BOUND_DURATION, UPPER_BOUND_SEGMENTS,
-    PRIMITIVE, BB1, SK1,
-    WAMF1,
-    CORPSE,
-    CORPSE_IN_SK1,
-    CORPSE_IN_BB1,
-    SCROFULOUS,
-    CORPSE_IN_SCROFULOUS)
+UPPER_BOUND_RABI_RATE = 1e10
+"""Maximum allowed rabi rate
+"""
 
-from .predefined import new_predefined_driven_control
+UPPER_BOUND_DETUNING_RATE = UPPER_BOUND_RABI_RATE
+"""Maximum allowed detuning rate
+"""
 
-__all__ = ['BB1',
-           'CORPSE',
-           'CORPSE_IN_BB1',
-           'CORPSE_IN_SCROFULOUS',
-           'CORPSE_IN_SK1',
-           'LOWER_BOUND_DURATION',
-           'PRIMITIVE',
-           'SCROFULOUS',
-           'SK1',
-           'UPPER_BOUND_DETUNING_RATE',
-           'UPPER_BOUND_DURATION',
-           'UPPER_BOUND_RABI_RATE',
-           'UPPER_BOUND_SEGMENTS',
-           'WAMF1',
-           'DrivenControl',
-           'new_predefined_driven_control']
+UPPER_BOUND_DURATION = 1e6
+"""Maximum allowed duration of a control
+"""
+
+LOWER_BOUND_DURATION = 1e-12
+"""Minimum allowed duration of a control
+"""
+
+UPPER_BOUND_SEGMENTS = 10000
+"""Maximum number of segments allowed in a control
+"""
+
+##### Types of driven controls ######
+
+PRIMITIVE = 'primitive'
+"""Primitive control
+"""
+
+BB1 = 'BB1'
+"""First-order Wimperis control, also known as BB1
+"""
+
+SK1 = 'SK1'
+"""First-order Solovay-Kitaev control
+"""
+
+WAMF1 = 'WAMF1'
+"""First-order Walsh sequence control
+"""
+
+CORPSE = 'CORPSE'
+"""Dynamically corrected control - Compensating for Off-Resonance with a Pulse Sequence (COPRSE)
+"""
+
+CORPSE_IN_BB1 = 'CORPSE in BB1'
+"""Concatenated dynamically corrected control - BB1 inside COPRSE
+"""
+
+CORPSE_IN_SK1 = 'CORPSE in SK1'
+"""Concatenated dynamically corrected control - First order Solovay-Kitaev inside COPRSE
+"""
+
+SCROFULOUS = 'SCROFULOUS'
+"""Dynamically corrected control -
+   Short Composite Rotation For Undoing Length Over and Under Shoot (SCROFULOUS)
+"""
+
+CORPSE_IN_SCROFULOUS = 'CORPSE in SCROFULOUS'
+"""Concatenated dynamically corrected control - CORPSE inside SCROFULOUS
+"""
