+++ qctrlopencontrols/driven_controls/driven_control.py
@@ -20,13 +20,13 @@ driven_controls.driven_controls
 import json
 import numpy as np
 
-from qctrlopencontrols.exceptions import ArgumentsValueError
-from qctrlopencontrols.base import create_repr_from_attributes
+from ..exceptions.exceptions import ArgumentsValueError
+from ..base.utils import create_repr_from_attributes
 
-from qctrlopencontrols.globals import (
+from ..globals import (
     QCTRL_EXPANDED, CSV, JSON, CARTESIAN, CYLINDRICAL)
 
-from .constants import (
+from ..driven_controls import (
     UPPER_BOUND_SEGMENTS, UPPER_BOUND_RABI_RATE, UPPER_BOUND_DETUNING_RATE,
     UPPER_BOUND_DURATION, LOWER_BOUND_DURATION)
