+++ qctrlopencontrols/driven_controls/predefined.py
@@ -23,17 +23,12 @@ can be found at https://docs.q-ctrl.com/control-library
 
 import numpy as np
 
-from qctrlopencontrols.exceptions import ArgumentsValueError
-from .driven_control import DrivenControl
+from ..exceptions.exceptions import ArgumentsValueError
+from ..driven_controls import (
+    BB1, CORPSE, CORPSE_IN_BB1, CORPSE_IN_SCROFULOUS, CORPSE_IN_SK1,
+    PRIMITIVE, SCROFULOUS, SK1, WAMF1)
 
-from .constants import (
-    PRIMITIVE, BB1, SK1,
-    WAMF1,
-    CORPSE,
-    CORPSE_IN_SK1,
-    CORPSE_IN_BB1,
-    SCROFULOUS,
-    CORPSE_IN_SCROFULOUS)
+from .driven_control import DrivenControl
 
 
 def new_predefined_driven_control(
