+++ qctrlopencontrols/dynamic_decoupling_sequences/__init__.py
@@ -13,31 +13,58 @@
 # limitations under the License.
 
 """
-================
-Sequences module
-================
+============================
+dynamic_decoupling_sequences
+============================
 """
 
-from .constants import (
-    UPPER_BOUND_OFFSETS, SPIN_ECHO, CARR_PURCELL,
-    CARR_PURCELL_MEIBOOM_GILL, UHRIG_SINGLE_AXIS,
-    PERIODIC_SINGLE_AXIS, WALSH_SINGLE_AXIS, QUADRATIC,
-    X_CONCATENATED, XY_CONCATENATED)
+UPPER_BOUND_OFFSETS = 10000
+"""Maximum number of offsets allowed in a Dynamical
+Decoupling sequence.
+"""
 
-from .dynamic_decoupling_sequence import DynamicDecouplingSequence
-from .predefined import new_predefined_dds
-from .driven_controls import convert_dds_to_driven_control
+MATPLOTLIB = 'matplotlib'
+"""Matplotlib format of data for plotting
+"""
 
-__all__ = ['CARR_PURCELL',
-           'CARR_PURCELL_MEIBOOM_GILL',
-           'UPPER_BOUND_OFFSETS',
-           'PERIODIC_SINGLE_AXIS',
-           'QUADRATIC',
-           'SPIN_ECHO',
-           'UHRIG_SINGLE_AXIS',
-           'WALSH_SINGLE_AXIS',
-           'X_CONCATENATED',
-           'XY_CONCATENATED',
-           'DynamicDecouplingSequence',
-           'convert_dds_to_driven_control',
-           'new_predefined_dds']
+###### Types of Dynamic Decoupling Sequences #######
+
+RAMSEY = 'Ramsey'
+"""Ramsey sequence
+"""
+
+SPIN_ECHO = 'spin echo'
+"""Spin echo (SE) dynamical decoupling sequence
+"""
+
+CARR_PURCELL = 'Carr-Purcell'
+"""Carr-Purcell (CP) dynamical decoupling sequence
+"""
+
+CARR_PURCELL_MEIBOOM_GILL = 'Carr-Purcell-Meiboom-Gill'
+"""Carr-Purcell-Meiboom-Gill (CPMG) dynamical decoupling sequence
+"""
+
+UHRIG_SINGLE_AXIS = 'Uhrig single-axis'
+"""Uhrig (single-axis) dynamical decoupling sequence
+"""
+
+PERIODIC_SINGLE_AXIS = 'periodic single-axis'
+"""Periodical dynamical decoupling sequence
+"""
+
+WALSH_SINGLE_AXIS = 'Walsh single-axis'
+"""Walsh dynamical decoupling sequence
+"""
+
+QUADRATIC = 'quadratic'
+"""Quadratic dynamical decoupling sequence
+"""
+
+X_CONCATENATED = 'X concatenated'
+"""X-Concatenated dynamical decoupling sequence
+"""
+
+XY_CONCATENATED = 'XY concatenated'
+"""XY-Concatenated dynamical decoupling sequence
+"""
