+++ qctrlopencontrols/exceptions/__init__.py
@@ -11,13 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-=================
-Exceptions module
-=================
-"""
-
-from .exceptions import ArgumentsValueError
-
-__all__ = ['ArgumentsValueError']
