+++ qctrlopencontrols/qiskit/__init__.py
@@ -11,13 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-=============
-qiskit module
-=============
-"""
-
-from .quantum_circuit import (convert_dds_to_qiskit_quantum_circuit)
-
-__all__ = ['convert_dds_to_qiskit_quantum_circuit']
