+++ tests/test_driven_controls.py
@@ -21,10 +21,10 @@ import os
 import numpy as np
 import pytest
 
-from qctrlopencontrols.exceptions import ArgumentsValueError
+from qctrlopencontrols.exceptions.exceptions import ArgumentsValueError
 from qctrlopencontrols import DrivenControl
 
-from qctrlopencontrols import (
+from qctrlopencontrols.driven_controls import (
     UPPER_BOUND_SEGMENTS, UPPER_BOUND_RABI_RATE, UPPER_BOUND_DETUNING_RATE)
